requirements.txt
python-pptx
yfinance
openpyxl
python-dotenv
boto3